refactor: 02_vector_db.py 를 본래 의도(vector DB 빌드)로 복원, build_vector_db.py 통합
이전엔 02_vector_db.py 가 graph DB 추출 코드로 repurpose 돼있었음.
03_graph_db.py 와 거의 같은 함수가 두 곳에 있고 02 는 옛 버전이라
PR #18/#25/#30 의 fix (notices 처리, isinstance 체크, _NULL_LIKE_NAMES,
traceback 로깅, summarize_document, cp949 stdout 등) 가 누락됨.

→ 02 를 실수로 실행하면 clear_db() 로 graph 삭제 + 옛 코드로 빌드되어
  graph 품질 망가짐. 위험.

변경:
- pipeline/02_vector_db.py: dead graph 빌드 코드 제거, vector DB 빌드
  스크립트로 대체 (chunk_size=500 / overlap=50 / RecursiveCharacterTextSplitter
  / Upstage embedding-passage / Chroma knu_cse_upstage_pro)
- pipeline/build_vector_db.py 제거: 02_vector_db.py 와 중복.
  README/evaluate_all_models.py 의 안내 ("02_vector_db.py 실행해서 vector
  DB 빌드") 와도 일치하게 됨.
